pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ==============================================================================
# Métadonnées du projet
# ==============================================================================

[project]
name = "Prométhée"
version = "3.0.0"
description = "Assistant IA web multi-utilisateurs avec outils intégrés, RAG, mémoire long terme et support Légifrance/Judilibre"
readme = "README.md"
license = { text = "AGPL-3.0-only" }
requires-python = ">=3.10"

keywords = ["ai", "llm", "rag", "assistant", "legifrance", "fastapi"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# ── Dépendances obligatoires ──────────────────────────────────────────────────
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
]

# ── Dépendances optionnelles (groupes) ────────────────────────────────────────
[project.optional-dependencies]

# Outils web
web = [
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",
    "lxml>=5.0.0",
]

# Analyse de données
data = [
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "numpy>=1.26.0",
]

# RAG — base vectorielle
rag = [
    "qdrant-client>=1.9.0",
    "numpy>=1.26.0",
]

# PDF et OCR
ocr = [
    "PyMuPDF>=1.24.0",
    "Pillow>=10.3.0",
    "pytesseract>=0.3.10",
]

# Documents Office + export PDF
office = [
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "reportlab>=4.1.0",
    "docx2txt>=0.8",
]

# Rendu Markdown et coloration syntaxique (interface chat)
rendering = [
    "markdown>=3.5.0",
    "Pygments>=2.17.0",
    "tiktoken>=0.7.0",
]

# Serveur FastAPI multi-utilisateurs
server = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
]

# Grist (tableur collaboratif)
grist = [
    "requests>=2.31.0",
]

# Audio
audio = [
    "pydub>=0.25.1",
]

# Modèles locaux Ollama
local = [
]

# SQL (drivers additionnels — installés à la demande via python_install)
sql = [
    "psycopg2-binary>=2.9.0",   # PostgreSQL
    "pymysql>=1.1.0",            # MySQL / MariaDB
]

# Tout (sauf drivers SQL et audio)
all = [
    "promethee-ai[web,data,rag,ocr,office,rendering,grist,local,server]",
]

# Développement
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "coverage>=7.5.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

# ── Point d'entrée ────────────────────────────────────────────────────────────
[project.scripts]
promethee = "main:__main__"
promethee-server = "server.main:app"

# ==============================================================================
# Outils
# ==============================================================================

# ── Ruff (linter + formateur) ─────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100
src = [".", "core", "tools", "server"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (géré par le formateur)
    "B008",   # do not perform function calls in default arguments
    "SIM108", # ternary operator (parfois moins lisible)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S102"]  # assert et exec autorisés dans les tests

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true   # stubs tiers non exhaustifs
exclude = ["tests/"]    # aucune UI Qt à exclure

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["core", "tools"]
omit = [
    "tests/*",
    "*/site-packages/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ── Setuptools ────────────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "tools*", "server*"]
exclude = ["tests*", "assets*", "frontend*"]
